It's now possible to refresh the lists in the drawer to get all lists who belong to you Fixed a bug in the Filemanager, when the file for the list didn't exist

+++ lib/Models/ShoppingList.dart
@@ -25,7 +25,8 @@ class ShoppingList {
 
   Future save() async {
     var filename = "ShoppingLists/${id.toString()}.sl";
-    await FileManager.deleteFile(filename);
+    if (FileManager.fileExists(filename))
+      await FileManager.deleteFile(filename);
     await FileManager.createFile(filename);
     await FileManager.writeln(filename, name);
